action/import: Detect a snapshot ID while creating a linked cone
When a new VM is creating as a linked clone, we detect the current snapshot ID to pass it to the clone command.
If there is no snapshot yet, the new one will be created.
It is needed to avoid a lot of useless snapshots in the box, because by default `prlctl clone --linked` always initiates a new snapshot creation.