Use temporary lock file for boxes to prevent early unregister in the race
In parallel run of multi-machine environment the same box image could be cloning
simultaneously by more then one envs. So, the box should be unregister by the latest env.

To prevent a race condition here, we add a temporary lease file with a counter.
It is incremented for incoming envs (new clones) and decremented after the cloning.
The last env removes the lock file and unregisters the box.

Fixes GH-243