Make `start` action run provisioners if VM is running
That is a port of the upstream Vagrant commit:
https://github.com/mitchellh/vagrant/commit/e42f346b1d5891b5bd3651564894b95ed3b8b6cc

Previously, there was no one gesture that would start a VM if it was not
running and run the appropriate provisioners regardless of its original
state. `vagrant up` did nothing if the VM was running, while
`vagrant provision` did nothing if the VM was not running.

Change the semantics of `vagrant up`, via the start actions of the providers,
to go through the provisioning logic even if the VM is already running.
The semantics of `run: "once"` vs `run: "always"` are respected.