gemspec: Fix compatibility with Ruby 2.7 and older
Vagrant 2.3.* is still using Ruby 2.7 as the embedded ruby interpreter. We should not include a strict "required_ruby_version"
in our gemfile, otherwise it won't be possible to install our plugin on Vagrant 2.3* or older.
It's better to remove the line completely and rely on the compatibility defined by the upstream Vagrant.